+++ flutter_inappwebview_forge/lib/src/in_app_webview/main.dart
@@ -1,6 +1,7 @@
 export 'in_app_webview.dart';
 export 'in_app_webview_controller.dart';
 export 'headless_in_app_webview.dart';
+export 'in_app_webview_preloader.dart';
 export 'android/main.dart';
 export 'apple/main.dart';
 export '../find_interaction/find_interaction_controller.dart';